Fix md5 match checking in python3.
This should fix the other failing CI test, and is a real bug. In
python3, b64encode returns a bytes type instead of a string, which we
compare to a string. Comparing a bytes and a string always returns
False. We need to convert the bytes to a string ourselves.