Fix flipping into edit mode on dashboard.
It was super slow before. This ensures we don't remount shit when we
don't need to. But has the sad side effect that grid panels are always
draggable, even when not in edit mode. We need to patch the library to
fix it.
e Please enter the commit message for your changes. Lines starting