Fixes wandb/core#1452
Summaries were only being pushed to the file stream API on file modification. If watchdog never fires a modification event, maybe because it or the OS is overloaded, `wandb-summary.json` will be uploaded as a plain file, but not sent to the streaming API. Gorilla only watches for it to be streamed, so it won't populate summary metrics if it just gets uploaded as a plain file.

Fixed in a couple of ways:
- Stream the summary in `FileEventHandlerSummary.finish()` so that Gorilla always gets the final state of the summary to put in the metrics table.
- Walk the run directory immediately before ending file syncing so we can be almost certain that we've seen every single file in there. This has been a longstanding issue (with watchdog's unreliable filesystem polling) so I figured now was a good time to fix it anyway.