Fix repr(), .keys(), and .items() for HTTPSummary (#552)
Because of its lazy loading, HTTPSummary previously only showed items in
repr(), that had already been accessed by the user. This change makes it
always use the set of keys in the underlying JSON dict. It uses the
placeholder '...' for h5 objects that haven't been loaded yet because
they may be extremely large.

https://github.com/wandb/core/issues/2705