Fix deadlock.
Trying to track the final file upload to finish prepare when it's done
create a race and deadlock. Since two upload jobs can run in parallel
the one that thinks it's last may shutdown step_prepare before the other
job has done a prepare.

We don't need this behavior anyway, we can just rely on a single final
prepare. The behavior was there for an earlier version of the code
that needed a faster path for shutting down prepare.