Fix an issue where files sync'd to the wrong run. (#1092)
* Fix an issue where files sync'd to the wrong run.

This would happen if you had two runs running in parallel with wandb version 0.9.0. UploadJob switched to calling upload_urls rather than push on the internal api. upload_urls defaulted to using the run set in settings which was always the string "latest". This is super legacy and we don't want it.

This switches upload_urls to use current_run_id instead of 'latest' when no run is provided, and gets rid of other uses of latest

* Fix other cases that relied on settings['run']

* Doc updates

* update CHANGELOG

* Update doc strings and changelog

Co-authored-by: shawnlewis <shawnlewis@users.noreply.github.com>
Co-authored-by: Chris Van Pelt <vanpelt@gmail.com>