[Bugfix] Update wandb image for syncing  (#41)
Removes image spriting for images in image sequences. There is a few benefits here:
The old cli would override the first image in the array with the sprited version. This is incompatible with the new cling flow where the file path ends up getting updated later in bind_to_run.
Increases performance in the UI when we are only loading one index
Simpler implementation and more consistent with our other media types
Consumes less system resources on the host machine
Allows us to add future support to the UI for non heterogenous image sizes(In-demand feature)
Adds the step to history encoding
Ensures the data types are encoded as a history object before the summary so they are called by bind_to_run with their step instead of summary.
Adds a _datatype_callback to media files that are not is_tmp e.g. path objects
Image fix that was merged to the old CLI last week that properly handles image files(wandb/client#1152)