fix feature importance sorting
on the models with the coef_ attribute the feature importance array is returned as an (n_classes, n_features) array instead of an (n_features,) array. this breaks the assumption in the subsequent code that `importances` has shape (n_features,). This commit fixes that by taking the class-wise median of the importances in those instances to give a summary statistic for feature importances that applies across all classes