Gracefully handle string labels in wandb.sklearn.plot.classifier.calibration_curve
This patch addresses a bug that would cause a `wandb.sklearn.plot_classifier` call
to error out during the calibration curve generation if a list of string labels is
provided (the element-wise comparison fails in this case and returns a scalar
boolean, which caused an `AttributeError` exception to be raised. Instead of
raising, we now warn that this plot is being skipped.