Implement support for IPUs
This adds support for automatically logging various metrics of an IPU, such as temperature, clock speed and utilisation. This is implemented similarly to how TPU support is implemented, where it only runs if the necessary imports are available, in this case a package from our sdk, `gcipuinfo`. This is extracted into it's own class, `IPUProfiler`.

All stats that are returned by gcipuinfo are logged initially, but a lot are static so only a select few are logged on every invocation.

Some unit tests are added for `IPUProfiler` that mock the interface of the `gcipuinfo` package, as well as some unit tests for the `SystemStats` class to check that it interacts with the IPUProfiler class correctly. Note that no tests actually run anything on the IPU.