fix(launch): correct registry uri for vertex
build.py previously contained two functions for
construction image uris. The one that was being tested
was not the one being used in the vertex runner.
I deleted the function in use and replaced it with the
function we already test, construct_gcp_image_uri

construct_gcp_image_uri also added the tag to the URI
We add the tag to the uri when we push anyway so there
were two tags at the end of the uri string.
I removed the addition of the tag from construct_gcp_image_uri
and updated the test accordingly