refactor(artifacts): internally use Future to communicate success/failure of commit, not threading.Event (#4859)
(This PR is a redo of part of the reverted #4677)

Currently `ArtifactSaver` uses a `threading.Event` to wait for the artifact-commit to complete. Problem: if the artifact _fails_ to commit for any reason, then the ArtifactSaver hangs forever (blocking the entire SenderThread forever, too!).

This PR replaces the `Event` with a `concurrent.futures.Future`, which can communicate successful completion _or failure_. If the artifact fails to commit (e.g. because of a non-retryable error in `commitArtifact`, or because one of the files failed to upload), then the `Future` will still get resolved -- it will just get resolved with an Exception that will get logged and propagated back to the user-process.