fix(sdk): fix wandb tfevent sync issue (GH-3572)
Currently, if self._force is set on TBWatcher, we will attempt to parse tfevents from all files -- even those which aren't tfevents. This causes `wandb sync --sync-tensorboard` to fail silently on directories containing more than just tfevent files.

As the desired behaviour of `self._force` seems to be 'parse all tfevent files regardless of their hostname / start time', we patch the check in TBWatcher such that it only attempts to parse a file if it's actually a tfevent file.