Update import_hook.py with latest changes in the wrapt repository [draft] (#5292)
* fix(sdk): fix typo in comment in import_hooks.py file

This change was added to the wrapt repo as git commit f2e46f94 by
Tim Gates (@timgates42) in Dec 2020.

* fix(sdk): use new Python module import system hooks to eliminate deprecation warnings.

This change was added to the wrapt repo as git commit 53323ba4 by
Graham Dumpleton (@GrahamDumpleton) in Mar 2022.

* fix(sdk): Deal with module importers that don't implement newer API

This change was added to the wrapt repo as git commit df0e62c by
Graham Dumpleton (@GrahamDumpleton) in Mar 2022.

* fix(sdk): Fix threading issues with register_post_import_hook

This change was added to the wrapt repo as git commit c5173c9 by
Eugene Toder (@eltoder) in Aug 2022.

Upstream commit message:

1. Add a missing lock to ImportHookFinder.find_spec(). It should have
the same lock as find_module() since it accesses _post_import_hooks.

2. Fix a deadlock between notify_module_loaded() and ImportHookFinder.
This occurs because the former is calling import hooks under
_post_import_hooks_lock. If a hook is making any imports itself while
another thread is also making imports, the thread can grab the import
lock and wait for _post_import_hooks_lock in ImportHookFinder, while the
hook already has _post_import_hooks_lock and is waiting for the import
lock. This results in a deadlock. The solution is to call the hooks
outside of _post_import_hooks_lock.

Same for register_post_import_hook() and  ImportHookFinder.

3. Also fix an issue if register_post_import_hook() was called after a
module was already imported but was later removed from sys.modules. This
also makes the logic simpler.

* fix(sdk): Restore original loader after import

This change was added to the wrapt repo as git commit 2d9fd85 by
Eugene Toder (@eltoder) in Dec 2022 to fix upstream issue 221
("post_import_hooks break pkg_resources in wrapt 1.14").

* fix(sdk): Add extra check on whether __loader__ attribute exists.

This change was added to the wrapt repo as git commit 9bab802 by
Graham Dumpleton (@GrahamDumpleton) in Jan 2023.