refactor(integrations): Refactor autologging resolvers for more generic autolog monkey patch to accommodate requests that come from args/kwargs and adjust response to allow for a dictionary of wandb allowable media types or metrics for various packages or APIs.
This commit refactors the autologging classes PatchLLMAPI and AutologLLMAPI to PatchAPI and AutologAPI, respectively. Additionally, the argument resolver protocol RequestResponseResolver is changed to ArgumentResponseResolver. The request parameter is changed to args and kwargs, and the response parameter is changed to a dictionary of wandb media objects or metrics. This makes the code more concise and easier to understand.