fix(core): Don't retry when trying to upload a directory
Fixes WB-17558: `run.save(directory)` will retry until it runs out of retries.

This happened because of retryablehttp retries Go errors from the HTTP transport (i.e. not HTTP response error statuses):

https://github.com/hashicorp/go-retryablehttp/blob/main/client.go#L506-L507

Our request body is a ReadSeeker, so retryablehttp doesn't try to read the entire file into memory when we create a Request. Instead, it passes the ReadSeeker to the underlying transport in Do(). When Go's http library calls Read(), it gets an error saying that the file is a directory; then retryablehttp treats it as a retryable error!

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7225