refactor(core): Extract FileStream interface for testability
Description
---

Makes `FileStream` an interface with a private `fileStream` implementation so that a fake can be created when testing code that depends on `FileStream`.

I moved nil checks out of FileStream and into Sender, which is the only code where there can be a nil FileStream.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7309