refactor(core): Create a new watcher wrapper package
Description
---

Factors out the file-watching code out of the `runfiles` package to improve testability. I use this in the next PR to write tests for a bug fix in the uploader (it passes a relative path to the watcher, but it should be passing an absolute path).

Fun fact: I uncovered a race condition in the watcher package we use because of this!

Fun fact #2: I realized that comparing `mtime` like the watcher package does is wrong! https://apenwarr.ca/log/20181113

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7344