feat(core): Avoid race condition when uploading run files
Description
---
Fixes WB-18104.

When using "live" mode, or when `run.save()` is called repeatedly while a file is constantly changing, there can be multiple simultaneous uploads to the same file. The semantics of this depend on the server receiving the uploads, but for run files this is a race condition, and the final uploaded file version might not be the newest one.

This adds buffering behavior to `run.save()` when using `wandb-core`: if the process knows it's already uploading a given run file, it waits for that to complete before uploading that file again. If there are multiple upload requests during a single upload, the last one wins.

Additionally, this removes unreliable ModTime checks. Now we accidentally fail to upload a new version of a file when requested, but a live file will be uploaded at least twice: once immediately and once at the end. Fixing this is for another PR.

Testing
---
There is a unit test for the added code.

It's hard to manually test a race condition: I just tested that file uploads still work as expected.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7346