refactor(core): Remove overflow wait-skipping logic in collector.go
Description
---
The code was doing this before, which wasn't obvious:

1. Read from `input` until we process a max number of items or a timer expires
2. If we hit the max, then next time, read *a random number* of buffered items from `input` (still up to the max)

(2) is surprising, but it happens because Go's `select` statement chooses a random case if multiple channels are ready. When `readMore()` was called with `isOverflow=true`, then `delayChan` was a channel with 1 buffered event. If `input` had N<max buffered events, the method did *not* read all N of them, but rather read a random number between 0 and N.

There's a way to make it deterministic, but this behavior doesn't add value as far as I can tell.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7507