refactor(core): Don't assume file upload URL doesn't change
Description
---
The query part of the URL can change, e.g. if it contains an expiration time or an auth token.

There was no bug here as far as I can tell, but there was certainly the opportunity for a bug. I rearranged the code to make the assumption unnecessary.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7515