refactor(core): Move (nil, nil) retryablehttp output handling
Description
---
Moves the err==nil && resp==nil check into `api` to avoid duplicating it. We had this check in `filetransfer` but we also need it to apply in `filestream`.

As I recall, this is some bug in the retryablehttp client, but I don't remember if we have a tracking bug or a way to reproduce it.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7512