feat(core): Don't panic in filestream
Description
---
Makes `filestream` react gracefully to errors that prevent further uploads. After a fatal error, `filestream` operations become non-blocking no-ops.

The reason we do this rather than killing the whole process is because we want to be very careful about data loss. Shutting down just the `filestream` allows the rest of the process to continue recording transactions to disk, making it possible to `wandb sync` at a later time without interrupting a user's experiment.

Normally, this pattern should not be used: a panic means the program is in an undefined state and it is not safe to continue. In our case, `filestream` does not touch any state outside of itself, so a panic doesn't need to break the rest of the process. But note that **this is a dangerous pattern** and we could end up corrupting data in the future if that assumption is broken.

Two changes enable this:

1. All panics are replaced by explicit errors
2. All goroutines started by filestream are guarded by a `recover()`

(2) is a catch-all mechanism to gracefully shut down just the filestream component for _unexpected_ panics caused by programming errors. This can never happen because we write bug-free code, but I add it anyway as a matter of principle.

Reviewers: raubitsj, dmitryduev

Reviewed By: dmitryduev

Pull Request: https://github.com/wandb/wandb/pull/7510