feat(core): Print error in terminal if filestream breaks
Description
---

Completes WB-18452 (main fix in https://github.com/wandb/wandb/pull/7510, but this makes it user-friendly).

If the filestream shuts down, let the user know to run `wandb sync` manually. NOTE: the message is printed at the time of the fatal error, not at the end of the run, so it could get lost if there are many console messages. Adding this is a bit tougher with what we have now, so I'm going to do it in another PR.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7517