refactor(core): Make `filestream.Update` deps more explicit
Description
---
Instead of accepting the private `fileStream` struct, make `Update.Apply()` operate on an `UpdateContext`, which is just an explicit collection of capabilities. An `Update` can't do arbitrary things to the filestream---this makes that clear.

This paves the way for improving code cohesion. Instead of having `updatehistory.go` under `filestream`, we could put it under `runhistory` to make it easy to follow how history data is processed. Likewise for summary and so on.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7568