chore(sdk): Update TBRecord docs
Description
---
Further refine the documentation for TBRecord now that I've stared at the Python implementation for a couple of weeks.

I was hoping to force `log_dir` and `root_dir` to be absolute paths for simplicity, but after updating `interface.py` to make the paths absolute and testing manually, I observed that it changed behavior under the orig. impl., so I undid the changes. Should be possible to simplify these semantics once we delete the Python code.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7746