fix(core): Correctly sample history metrics with nested keys
Description
---
The sampler implementation in `handler.go` was not reading the `NestedKey` field of `HistoryItem`. This pulls out history sampling logic into a dedicated `runhistorysampler.go` file and implements it properly.

There is a bug in Python where it doesn't _read_ the `nested_key` field. I left a comment about it in `runhistorysampler.go`. Amazingly, this is countered by a bug in the Python TensorBoard integration, which seems to produce items that use `key` instead of `nested_key`, causing them to have sparklines anyway.

Testing
---
The following doesn't print a sparkline with the Python implementation but does with `wandb.require("core")`:

```python
with wandb.init() as run:
    for x in range(100):
        run.log({"my": {"value": x}})
```

The TensorBoard integration with and without core prints sparklines.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7768