chore(core): Utility for line-oriented files
Description
---
Creates `linefile.go` in the `runconsolelogs` package for writing the `output.log` file.

`linefile.go` implements routines to efficiently modify the last few lines of a file, which is useful for terminal emulation.

An alternative is to keep an in-memory map from line numbers to lines and a "finalized" flag for each line, and to only write finalized lines to the file. Properly managing the memory for this is more complicated than just updating a file, IMO.

Testing
---
Unit tests and manual testing in the final PR in this chain that performs terminal emulation for captured stdout/stderr streams.

Reviewers: dmitryduev

Reviewed By: dmitryduev

Pull Request: https://github.com/wandb/wandb/pull/7824