refactor(core): Define FileStreamRequest, delete CollectorState
Description
---
Replaces "CollectorState" / "FsTransmitData" and cleans up that whole mess.

To review, read `filestreamrequest.go` where most of the work happens.

I define following types:

* `FileStreamRequest`
* `FileStreamRequestJSON`
* `FileStreamRequestReader`
* `FileStreamState`

An important property of filestream is that any sequence of requests can be compressed into one request. This wasn't well represented by the old abstractions. Now, there is a `FileStreamRequest` type with a `Merge()` method.

The `FileStreamRequest` represents a request for an idealized API; on the other hand, `FileStreamRequestJSON` is the reality we live in:

* Some requests can't _actually_ be combined. The filestream API doesn't allow us to update nonconsecutive lines in a file in one request, which happens when we perform terminal emulation.
* When that happens, we break the idealized request into multiple smaller requests. But in doing so, we must preserve an ordering guarantee: the run's exit code must only be set in the final request.

`FileStreamRequestReader`: The `FileStreamRequest` struct is well-suited for quick merge operations, but breaking it into a `FileStreamRequestJSON` "head" and a `FileStreamRequest` "tail" (containing any remaining data) requires memory copy operations. If we're unlucky, this could be a lot of work---like if we buffered a very large request. The Reader defers this work until we actually need it instead of repeating it each time we merge a request into the buffer. Concretely, we avoid unnecessary copying while waiting for rate limits (1 request / 15 sec) or the previous network operation to complete. Copying is necessary because request buffering must happen in a different goroutine than sending.

`FileStreamState`: This is state that would ideally exist on the backend. It's kept separate from the idealized `FileStreamRequest`, but we need it to create the real `FileStreamRequestJSON` values. It's stuff like a console line offset for appending logs to a resumed run, or the next history line number (since we always append to history).

Testing
---
It was easy to write thorough `FileStreamRequest` unit tests which speaks to its success as a model :)

Some manual testing as well:

```python
import time
import sys

from tqdm import tqdm, trange
import wandb

run = wandb.init()

for i in trange(60):
    # Exercise console logging. This forces tqdm to send out
    # ANSI escape sequences to the terminal.
    tqdm.write(f"Iteration {i}:")

    # Exercise history + summary and buffering behaviors:
    for j in trange(1000):
        run.log({"i": i, "j": j})
        # Wait to exercise system metrics logic and also
        # rate-limiting and buffering.
        time.sleep(1.0 / 1000.0)

sys.exit(1)
```

It's hard to manually test the UploadedFiles and Preempting parts. If you run with WANDB_CORE_DEBUG=true, you can see that all standard files appear in the "uploaded" field of a filestream request at some point:

```
jq '
  select(.msg|contains("filestream: post request")) |
  .request | fromjson |
  .uploaded
' wandb/debug-internal.log
```

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7879