chore(core): Add `@override` to method to silence lint
Description
---
We should be marking overridden methods with `@override` (introduced in PEP 698). This silences the Ruff's D102 lint (requiring public methods to have docstrings) because the method in the base class provides the docstring, and it also makes the intention more explicit.

The import is from `typing_extensions` rather than `typing` because `typing.override` was added in Python 3.12.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7892