chore(sdk): communicate_status => deliver_status
Description
---
Replaces `communicate_status()`, which uses the deprecated "communicate" verb and MessageRouter path, by `deliver_status()` which uses the newer Mailbox.

This changes a method in `streams.py` to wait indefinitely rather than with a 5s timeout. Users that previously hit the `AssertionError` in that method will now see either:

* A success, if the 5s timeout was just not enough (what we expect)
* A hanging process, if a response was never going to happen. This would be a bug.

We do not currently have a mechanism to propagate deadlines, but it would be more appropriate than using an indefinite timeout. The goal of this PR is to replace a deprecated codepath.

Testing
---
A simple `wandb.init()` script with `WANDB__REQUIRE_CORE=false` to trigger the `streams.py` change.

I grepped for "communicate_status" in the codebase to find all usages.

Reviewers: raubitsj

Reviewed By: raubitsj

Pull Request: https://github.com/wandb/wandb/pull/8125