refactor: replace mailbox.ContextCancelledError (#9365)
Replace `mailbox.ContextCancelledError` by `retry.RetryCancelledError`.

This error is only raised in `retry.py` and only caught in `sender.py`. It subclasses `mailbox.MailboxError`, but this is only caught in `wandb_run.py` while running `MailboxHandle.wait()` which cannot throw `ContextCancelledError`.

There was no reason for this to be in `mailbox.py`.