chore(launch): clean launch agent mock between launch kubernetes tests (#9360)
Description
-----------
<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->
- [Slack thread](https://weightsandbiases.slack.com/archives/C0108SLEP6K/p1738180653893699)

Some Launch-related tests are failing because the `LaunchAgent` is being instantiated before the test. This adds a fixture to clear the `LaunchAgent` before running the affected tests.

<!--
NEW: We're using a new changelog format that's more useful for users. Please
see CHANGELOG.unreleased.md for details and update on relevant changes such as feature
additions, bug fixes, or removals/deprecations.
-->
- [x] I updated CHANGELOG.unreleased.md, or it's not applicable


Testing
-------
The test is flaky so it's hard to say for sure, but it passes locally and in CI.

<!--
Ensure PR title compliance with the [conventional commits standards](https://github.com/wandb/wandb/blob/main/CONTRIBUTING.md#conventional-commits)
-->