refactor: don't create perf.log file when importing tools/perf/scripts (#9388)
The `get_logger` function was opening a "perf.log" file in the current working directory by instantiating a `logging.FileHandler`. This caused the "perf.log" file to appear when unrelated Python tools imported that directory in the background---I believe the culprit is `pytest`'s test discovery mechanism in my case, which VSCode runs automatically.