refactor: include backend error messages in operation stats (#9443)
When the HTTP response from the backend contains error messages in a known format, include them in operation stats, making `run.finish()` more informative and also `wandb.init()` (with PR #9431).

The logic is the same as in `wandb/util.py::parse_backend_error_messages()`, and the test cases are the same as those in `test_normalize.py`.