refactor: asyncio compatibility tools (#9379)
Adds `asyncio_compat.py` file with some important utilities for using `asyncio` safely.

In `asyncio`, cancelling a task does not cancel subtasks that it started. Python 3.11 adds a `TaskGroup` class to fix this, but we need custom code for Python 3.8 - 3.10.

Using `asyncio.run` inside an `asyncio` event loop is not allowed. In case any of our users call `wandb` functions in an `asyncio` context, the `asyncio_compat.run` function starts an `asyncio` loop in a new thread.