refactor: new mailbox implementation (#9380)
A reimplementation of `mailbox.py` with tests and documentation, that works with `asyncio`.

This PR does not delete the old mailbox, and no code uses the new mailbox yet.

## Design notes

The `MailboxProbe` / `MailboxProgress` mechanism in the old implementation relied heavily on callbacks and was undocumented / untyped making it difficult to use. The new mailbox provides an async interface for the same purpose, see PR #9382.

The `wait()` method on `MailboxHandle` is there for compatibility with old code, so that it's possible to migrate incrementally.

The old mailbox's "keepalive" mechanism is replaced by a `Mailbox.close()` method which the Router is meant to call when it stops reading new messages (in a `finally` block).

## test_new_mailbox.py

The test file is a little repetitive because it tests both the sync (`wait_or`) and async (`wait_async`) versions of the same functionality. Most tests are short and simple, but the `test_unblocks_wait_*` tests are more involved as they attempt to test concurrency.

I don't call it `test_mailbox.py` because `pytest` doesn't like when multiple test files have the same name (even in different directories). I'll rename it after deleting the old mailbox.