refactor: replace old mailbox with new mailbox (#9420)
Deletes the old mailbox implementation and updates all code to use the new one.

Nice side effect: the spinners shown in `run.finish()` now update 10 times per second instead of once every second. We still only poll once a second.

The new `MailboxHandle.wait` method has the same semantics as the old one, so I did not update it in this PR. This PR only updates code that used the old `on_probe`/`on_progress`/`wait_all` functionality.

The new mailbox does not have a "keepalive" flag. In the old mailbox, handles created with this flag would periodically generate KeepAlive requests in `wait()` and use this to detect when the connection to the service process is closed. Instead, the Router object that delivers responses to it calls `mailbox.close()` when it is done, which also happens if the connection to the internal service process closes.