refactor: remove unused communicate_login and publish_login (#9456)
Removes `communicate_login` and `publish_login` methods from `InterfaceShared` as they are not used anywhere, and the former relies on the old `communicate` mechanism.

Since these are not used and were never saved to any transaction logs (due to being local), I also removed all the corresponding handling code in `wandb-core` and `legacy-service`.