refactor: remove "process_check" from InterfaceShared (#9464)
The `_process` and `_process_check` fields are only used by `InterfaceQueue`, so this PR removes them from `InterfaceShared`.

`Backend` is the only place that instantiates `InterfaceQueue` passing the `process` parameter.

`InterfaceRelay` inherits from `InterfaceQueue` and is only instantiated in `streams.py` with `_process_check` set to False, so it doesn't need the `process` parameter at all.