refactor: raise MailboxClosedError when creating handle in closed mailbox (#9465)
Raise an error instead of allowing the creation of a handle to which a response will never be delivered.

This PR also updates `StreamRecord` in `legacy-service` to own its own `Mailbox` instead of sharing one from `StreamMux`. A `StreamRecord` corresponds to a Run and starts up a `MessageRouter` that delivers responses to the mailbox. When the run ends, the `MessageRouter` closes the mailbox, so it cannot be reused by future runs.

There was no need for StreamRecords to share a Mailbox because a Record delivered to a stream can only get a Result from that same stream. Using different Mailbox instances can never result in a situation where `require_response` is called on one `Mailbox` and the result is delivered to another.