fix: format specifier in WorkSentinel DebugInfo() (#9504)
Fixes the format specifier used in `WorkSentinel.DebugInfo()`. Using `%s` is incorrect when the argument (of interface type `any`) does not have the concrete type `string`. Specifically, it was producing the string

```
WorkSentinel(%!s(stream.senderSentinel=1))
```

Where `%!s` is `fmt.Sprintf`'s way of indicating that it caught and suppressed a panic. Now, with the added `String()` method on the `senderSentinel` type, the above becomes

```
WorkSentinel(senderSentinel(1))
```