refactor: change Mailbox to operate on ServerRequest / ServerResponse (#9488)
Generalize `Mailbox` to work on the `ServerRequest` / `ServerResponse` level instead of `Record` / `Result`.

`MailboxHandleT` is a generic class that replaces `MailboxHandle` to allow for both `MailboxHandleT[pb.Result]` and `MailboxHandleT[spb.ServerResponse]`. `MailboxHandle` is an alias to `MailboxHandleT[pb.Result]` to keep this PR small, but the next PR completes the renaming.

`Mailbox.deliver()` now requires `spb.ServerResponse` instead of `pb.Result`. The only callsite is in `MessageRouter`, which I updated. Unfortunately, `legacy-service` uses `MessageRouter` internally, so it has to continue to handle the `pb.Result` codepath. To do this, it creates an `spb.ServerResponse()` and sets its `result_communicate` to the given `pb.Result`.