refactor: remove `_router` from `InterfaceShared` (#9469)
Separate `MessageRouter` from `InterfaceShared`. A `MessageRouter` is a handle on a thread that delivers results from the service into a `Mailbox`.

This PR also updates `inform_attach` to use `Mailbox` and removes the deprecated `SockClient.send_and_recv` mechanism. This is necessary because `send_and_recv` assumes nothing else is reading on the socket, but this PR causes a reader thread to be created earlier in `wandb_init.py::_attach()`.

There can only be one `MessageSockRouter` per socket, and there is only one socket connection to the service per client process. If `InterfaceShared` owns `MessageSockRouter`, then there can only be one `InterfaceShared` instance at a time, and therefore only a single run at a time.

This PR makes code that instantiates `InterfaceShared` be responsible for managing a `MessageRouter`. There are only three such places:

* In `Backend`, when using the `x_disable_service` setting
* In `ServiceConnection`, in normal operation
    * The `ServiceConnection` owns a socket, so it also owns the router
    * The router thread lives as long as the socket is around, until `wandb.teardown()`
* In `streams.py` in `legacy-service`
    * Each `StreamRecord`, corresponding to an active run, owns an `InterfaceQueue` and a `MessageQueueRouter`