refactor: console capture redo (#9514)
Creates `console_capture.py` for patching `stderr` and `stdout`'s `write` methods.

The original way of doing this is in `redirect.py`, and I replace it in the next PR. `redirect.py` only allows a single hook at a time which forces us to only have a single run at a time. It also isn't thread-safe and doesn't document important edge cases (or much else).