refactor: allow multiple streams per connection in core (#9528)
Removes the `nc.stream` field and replaces it by calls to `streamMux.GetStream()`.

The risk is if we ever send messages without a stream ID to wandb-core, they will not be processed. `InterfaceSock` assigns a stream ID on all outgoing records; `ServiceConnection` in `service_connection.py` assigns a stream ID on Init, Finish, Attach and Start requests.

I have not checked other clients, but note that they would not work with legacy-service since it always requires the stream ID to be set.