fix: default reinit to "finish_previous" in notebooks (#9569)
Adds a `"default"` option to `reinit` that behaves like `"finish_previous"` in notebooks and `"return_previous"` elsewhere.

Originally broken in 0.19.0 in https://github.com/wandb/wandb/pull/8649. The type of `reinit` used to be incorrectly annotated as `bool` but initialized to `None`. After that PR rewrote Settings using pydantic, `reinit` started to be initialized to `False` instead of `None`. Then https://github.com/wandb/wandb/pull/9445 removed the `settings._jupyter` check in `wandb_init.py` because 

```
settings.reinit or (settings._jupyter and settings.reinit is not False)
```

was equivalent to `settings.reinit` since `reinit` could not be `None`.